% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter.R
\name{barycenter}
\alias{barycenter}
\title{Barycenter algorithm}
\usage{
barycenter(
  A,
  C,
  w,
  b_ext = NULL,
  barycenter_control = list(reg = 0.1, with_grad = FALSE, n_threads = 0, method = "auto",
    threshold = 0.1, max_iter = 1000, zero_tol = 1e-06, verbose = 0)
)
}
\arguments{
\item{A}{numeric matrix, source discrete densities (M * S)}

\item{C}{numeric matrix, cost matrix between source and target (M * N)}

\item{w}{numeric vector, weight vector (S)}

\item{b_ext}{numeric vector, only used to calculate quadratic loss
against the computed barycenter (default = NULL)}

\item{barycenter_control}{list, control parameters for the computation
\itemize{
\item reg: double, regularization parameter (default = .1)
\item with_grad: bool, whether to calculate the gradient w.r.t. A
\item n_threads: int, number of threads (only used for \code{method = "log"},
ignored by \code{method = "parallel"}, default = 0)
\item threshold: double, threshold value below which "auto" method
will default to \code{method = "log"} for stablized computation in log-domain
(default = .1)
\item max_iter: int, maximum iteration of \code{\link{barycenter}} algorithm
(default = 1000)
\item zero_tol: double, determine covergence (default = 1e-6)
\item verbose: int, print out debug info for the algorithm for every
\code{verbose} iteration (default to 0, i.e. not printing anything)
}}
}
\value{
list of results
\itemize{
\item b: numeric vector, computed barycenter
\item grad_A: gradient w.r.t. A (only with \code{with_grad = TRUE})
\item grad_w: gradient w.r.t. w (only with \code{with_grad = TRUE})
\item loss: double, quadratic loss between \code{b} and \code{b_ext}
(only with \code{with_grad = TRUE})
\item U, V: scaling variables for the Sinkhorn algorithm
(only with \code{method = "parallel"})
\item F, G: scaling variables for the Sinkhorn algorithm
(only with \code{method = "log"})
\item iter: iterations of the algorithm
\item err: condition for convergence
\item return_status: 0 (convergence), 1 (max iteration reached), 2 (other)
}
}
\description{
Barycenter algorithm to solve for entropy-regularized Optimal Transport
Barycenter problems.
}
\details{
This is the general function to solve OT Barycenter problem,
and it will use either parallel (\code{method = "parallel"}) or
log-stablized Barycenter algorithm (\code{method = "log"})
for solving the problem.
}
\examples{
A <- rbind(c(.3, .2), c(.2, .1), c(.1, .2), c(.1, .1), c(.3, .4))
C <- rbind(
  c(.1, .2, .3, .4, .5),
  c(.2, .3, .4, .3, .2),
  c(.4, .3, .2, .1, .2),
  c(.3, .2, .1, .2, .5),
  c(.5, .5, .4, .0, .2)
)
w <- c(.4, .6)
b <- c(.2, .2, .2, .2, .2)
reg <- .1

# simple barycenter example
sol <- barycenter(A, C, w, barycenter_control = list(reg = reg))

# you can also supply arguments to control the computation
# for example, including the loss and gradient w.r.t. `A`
sol <- barycenter(A, C, w, b, barycenter_control = list(reg = reg, with_grad = TRUE))

}
\references{
Peyré, G., & Cuturi, M. (2019). Computational Optimal Transport:
With Applications to Data Science.
\emph{Foundations and Trends® in Machine Learning}, 11(5–6), 355–607.
https://doi.org/10.1561/2200000073

Xie, F. (2025). Deriving the Gradients of Some Popular Optimal
Transport Algorithms (No. arXiv:2504.08722). \emph{arXiv}.
https://doi.org/10.48550/arXiv.2504.08722
}
\concept{Barycenter algorithms}
