% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_barycenter_args}
\alias{check_barycenter_args}
\title{used for \code{\link{barycenter()}} and \code{\link{wdl()}}}
\usage{
check_barycenter_args(brc_args)
}
\description{
used for \code{\link{barycenter()}} and \code{\link{wdl()}}
}
