% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cuda.R
\name{check_cuda}
\alias{check_cuda}
\title{Check if CUDA is available}
\usage{
check_cuda()
}
\value{
logical, TRUE if CUDA is available, FALSE otherwise
}
\description{
Check if CUDA is available for GPU computations.
}
\examples{
if (check_cuda()) {
  cat("CUDA is available for GPU computations.\n")
} else {
  cat("CUDA is not available.\n")
}
}
