% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headlines}
\alias{headlines}
\title{NYT headlines by economic policy uncertainty}
\format{
\subsection{headlines}{

A data frame with 18,507 rows and 2 columns:
\describe{
\item{headline}{Headline of the NYT News}
\item{ref_date}{Date of the News}
}
}
}
\source{
www.nytimes.com
}
\usage{
headlines
}
\description{
NYT headlines by economic policy uncertainty
}
\references{
Xie, F. (2020). Wasserstein index generation model: Automatic generation of
time-series index with application to economic policy uncertainty.
\emph{Economics Letters}, 186, 108874.
https://doi.org/10.1016/j.econlet.2019.108874
}
\keyword{datasets}
