% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvd.R
\name{tsvd}
\alias{tsvd}
\title{Truncated SVD}
\usage{
tsvd(M, k = 1, flip_sign = c("auto", "sklearn", "none"))
}
\arguments{
\item{M}{matrix, data to be analyzed}

\item{k}{int, number of columns/features to be kept}

\item{flip_sign}{character, one of the following: "auto", "sklearn", "none"}
}
\value{
matrix after dimension reduction
}
\description{
Truncated Singular Value Decomposition algorithm
}
\details{
Compute the truncated SVD for dimension reduction.
Note that SVD suffers from "sign indeterminacy," which means
that the signs of the output vectors could depend on the algorithm
and random state.
Two variants of "sign flipping methods" are implemented here,
one following the sklearn implementation on Truncated SVD,
another by Bro et al. (2008).
}
\references{
https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.TruncatedSVD.html

Bro, R., Acar, E., & Kolda, T. G. (2008).
Resolving the sign ambiguity in the singular value decomposition.
Journal of Chemometrics, 22(2), 135–140. https://doi.org/10.1002/cem.1122
}
\seealso{
\href{https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.TruncatedSVD.html}{Truncated SVD (sklearn)}
}
