% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdl.R
\name{wdl}
\alias{wdl}
\alias{wdl.character}
\alias{print.wdl}
\title{Wasserstein Dictionary Learning model}
\usage{
wdl(docs, ...)

\method{wdl}{character}(docs, specs = wdl_specs(), verbose = TRUE)

\method{print}{wdl}(object, topic = 0, token_per_topic = 5, ...)
}
\arguments{
\item{docs}{character vector, sentences to be analyzed}

\item{verbose}{bool, whether to print useful info}

\item{topic}{int, number of topic to be printed}

\item{token_per_topic}{int, number of tokens to be printed}

\item{spec}{list, model specification for the WDL
see \code{\link{wdl_spec()}} for reference}
}
\value{
topics and weights computed from the WDL given the input data
}
\description{
Wasserstein Dictionary Learning (WDL) model for topic modeling
}
\details{
This is the re-implementation of WDL model from ground up,
and it calls the \code{\link{barycenter()}} under the hood
(to be precise directly calling the underlying C++ routine
for \code{\link{barycenter()}})
}
\examples{
# simple WDL example
sentences <- c("this is a sentence", "this is another one")
wdl_fit <- wdl(sentences)


}
\references{
Peyré, G., & Cuturi, M. (2019). Computational Optimal Transport:
With Applications to Data Science.
\emph{Foundations and Trends® in Machine Learning}, 11(5–6), 355–607.
https://doi.org/10.1561/2200000073

Schmitz, M. A., Heitz, M., Bonneel, N., Ngolè, F., Coeurjolly, D.,
Cuturi, M., Peyré, G., & Starck, J.-L. (2018).
Wasserstein dictionary learning:
Optimal transport-based unsupervised nonlinear dictionary learning.
\emph{SIAM Journal on Imaging Sciences}, 11(1), 643–678.
https://doi.org/10.1137/17M1140431

Xie, F. (2025). Deriving the Gradients of Some Popular Optimal
Transport Algorithms (No. arXiv:2504.08722). \emph{arXiv}.
https://doi.org/10.48550/arXiv.2504.08722
}
