% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig-package.R
\docType{package}
\name{wig-package}
\alias{wig}
\alias{wig-package}
\title{wig: Wasserstein Index Generation model in R}
\description{
Implementing several Optimal Transport algorithms in Fangzhou Xie (2025) \doi{10.48550/arXiv.2504.08722} and re-implementing of Wasserstein Index Generation model by Fangzhou Xie (2020) \doi{10.1016/j.econlet.2019.108874}.
}
\author{
\strong{Maintainer}: Fangzhou Xie \email{fangzhou.xie@rutgers.edu}

}
\keyword{internal}
