% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headlines}
\alias{headlines}
\title{NYT headlines by economic policy uncertainty}
\format{
\subsection{headlines}{

A data frame with 18,507 rows and 2 columns:
\describe{
\item{headline}{Headline of the NYT News}
\item{ref_date}{Date of the News}
}
}
}
\source{
www.nytimes.com
}
\usage{
headlines
}
\description{
NYT headlines by economic policy uncertainty
}
\keyword{datasets}
