% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig.R
\name{wig}
\alias{wig}
\alias{wig.data.frame}
\alias{print.wig}
\alias{summary.wig}
\title{Wasserstein Index Generation model}
\usage{
wig(.data, date_col, docs_col, ...)

\method{wig}{data.frame}(.data, date_col, docs_col, specs = wig_specs(), verbose = TRUE, ...)

\method{print}{wig}(x, topic = 1, token_per_topic = 10, ...)

\method{summary}{wig}(object, topic = 1, token_per_topic = 10, ...)
}
\arguments{
\item{.data}{a dataframe containing the dates/datetimes and documents}

\item{date_col}{name of the column for dates/datetimes}

\item{docs_col}{name of the column for the texts/documents}

\item{...}{only for compatibility}

\item{specs}{list, model specification for WIG
see \code{\link{wig_specs}} for reference}

\item{verbose}{bool, whether to print useful info}

\item{x}{WIG model}

\item{topic}{int, number of topic to be printed}

\item{token_per_topic}{int, number of tokens to be printed}

\item{object}{WIG model}
}
\value{
"wig" class, i.e. list of the index and the WDL model
}
\description{
Wasserstein Index Generation (WIG) model for time-series sentiment index
autogeneration
}
\details{
This is the re-implementation of WIG model from scratch in R.
}
\examples{
# create a small dataset
wigdf <- data.frame(
  ref_date = as.Date(c("2012-01-01", "2012-02-01")),
  docs = c("this is a sentence", "this is another sentence"))

wigfit <- wig(wigdf, ref_date, docs,
  specs = wig_specs(wdl_control = list(num_topics = 2),word2vec_control = list(min_count = 1)),
  verbose = FALSE)

}
\references{
Xie, F. (2020). Wasserstein index generation model: Automatic generation of
time-series index with application to economic policy uncertainty.
\emph{Economics Letters}, 186, 108874.
https://doi.org/10.1016/j.econlet.2019.108874
}
\seealso{
\code{vignette("wdl-model")}
}
