% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig.R
\name{wig}
\alias{wig}
\alias{print.wig}
\title{Wasserstein Index Generation model}
\usage{
wig(.data, date_col, docs_col, ...)

\method{print}{wig}(object, topic = 1, token_per_topic = 5, ...)
}
\arguments{
\item{date_col}{name of the column for dates}

\item{docs_col}{name of the column for the texts}

\item{spec}{list, model specification for WIG
see \code{\link{wig_spec()}} for reference}

\item{verbose}{bool, whether to print useful info}
}
\description{
Wasserstein Index Generation (WIG) model for time-series sentiment index
autogeneration
}
\details{
This is the re-implementation of WIG model from scratch in R.
}
\examples{
# create a small dataset
wigdf <- data.frame(
  ref_date = as.Date(c("2012-01-01", "2012-02-01")),
  docs = c("this is a sentence", "this is another sentence")
)

wigfit <- wig(wigdf, ref_date, docs)



}
\references{
Xie, F. (2020). Wasserstein index generation model: Automatic generation of
time-series index with application to economic policy uncertainty.
\emph{Economics Letters}, 186, 108874.
https://doi.org/10.1016/j.econlet.2019.108874
}
